+++ src/test/java/racingcar/service/RaceResultServiceTest.java
@@ -3,7 +3,7 @@ package racingcar.service;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.dao.raceresult.RaceResultDao;
+import racingcar.dao.RaceResultDao;
 import racingcar.domain.RacingCars;
 import racingcar.entity.CarEntity;
 import racingcar.entity.RaceResultEntity;
